requirements.txt
torch>=2.4.0
pipecat-ai[silero]
websockets
sounddevice
numpy
loguru
protobuf
openai
